+++ policyengine_uk/parameters/gov/dfe/targeted_childcare_entitlement/max_income_tc_recipients.yaml
@@ -2,11 +2,11 @@ description: Families claiming the Tax Credits with income over this amount are 
 metadata:
   unit: currency-GBP  
   period: year
-  label: Income limit for tax credit recipients to access benefit-targeted free childcare
+  label: income limit for tax credit recipients to access benefit-targeted free childcare
   reference:
     - title: The Local Authority (Duty to Secure Early Years Provision Free of Charge) Regulations 2014, part 1.2.b
       href: https://www.legislation.gov.uk/uksi/2014/2147/regulation/1/made
     - title: 15 hours free education and childcare for 2-year-olds
       href: https://www.gov.uk/help-with-childcare-costs/free-childcare-2-year-olds-claim-benefits?step-by-step-nav=f237ec8e-e82c-4ffa-8fba-2a88a739783b
 values:
-  2015-01-01: 16_190
+  2014-01-01: 16_190
