+++ policyengine_uk/parameters/gov/dfe/targeted_childcare_entitlement/max_income_uc_recipients.yaml
@@ -2,11 +2,11 @@ description: The Department for Education sets this income threshold for Univers
 metadata:
   unit: currency-GBP 
   period: year
-  label: Income limit for universal credit recipients to access benefit-targeted free childcare
+  label: income limit for universal credit recipients to access benefit-targeted free childcare
   reference:
     - title: Free early education entitlement for two-year-olds under Universal Credit (Page 5)
       href: https://assets.publishing.service.gov.uk/media/5ab270baed915d4f2d097124/Government_response_FSM_and_EY_entitlements_under_Universal_Credit.pdf
     - title: 15 hours free education and childcare for 2-year-olds
       href: https://www.gov.uk/help-with-childcare-costs/free-childcare-2-year-olds-claim-benefits?step-by-step-nav=f237ec8e-e82c-4ffa-8fba-2a88a739783b
 values:
-  2015-01-01: 15_400
+  2014-01-01: 15_400
