+++ policyengine_uk/parameters/gov/dfe/targeted_childcare_entitlement/qualifying_benefits.yaml
@@ -7,9 +7,9 @@ metadata:
       href: https://consult.education.gov.uk/2-year-old-early-education-entitlement-team/free-early-education-for-disadvantaged-2-year-olds/supporting_documents/Eligibility%20for%20the%20free%20early%20education%20entitlement%20for%20twoyearolds%20from%20families%20with%20NRPF.pdf
   period: year
   unit: program
-  label: Qualifying benefits for targeted childcare entitlement
+  label: qualifying benefits for targeted childcare entitlement
 values:
-  2015-01-01:
+  2014-01-01:
     - jsa_income
     - esa_income
     - universal_credit
